requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
Brotli==1.1.0
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.4
mutagen==1.47.0
orjson==3.9.9
playwright==1.38.0
psycopg2-binary==2.9.9
pycryptodomex==3.19.0
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyee==9.0.4
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
spotipy==2.23.0
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.2
yt-dlp==2023.11.16
